Moving Characters to characters.lawful, in preparation for (now) the introduction of Vladimir and (not now) the introduction of a factory-style Character creation flow.
