+++ tests/characters/test_crypto_characters_and_their_powers.py
@@ -2,7 +2,7 @@ import eth_utils
 import pytest
 from constant_sorrow import constants
 
-from nucypher.characters import Alice, Character, Bob
+from nucypher.characters.lawful import Alice, Character, Bob
 from nucypher.crypto import api
 from nucypher.crypto.powers import CryptoPower, SigningPower, NoSigningPower, \
     BlockchainPower, PowerUpError
